+++ frontend/src/features/task/EditTaskDialog.tsx
@@ -49,6 +49,7 @@ import {
   selectAllLabels,
   selectLabelEntities
 } from "features/label/LabelSlice";
+import { formatDistanceToNow } from "date-fns";
 
 const mdParser = new MarkdownIt({ breaks: true });
 
@@ -64,8 +65,8 @@ const Main = styled.div`
 
 const Side = styled.div`
   margin-top: 2rem;
-  max-width: 200px;
-  min-width: 200px;
+  max-width: 220px;
+  min-width: 220px;
 `;
 
 const Header = styled.div`
@@ -134,6 +135,12 @@ const Description = styled.div`
 
 const DescriptionActions = styled.div`
   display: flex;
+`;
+
+const Text = styled.p`
+  color: #626e83;
+  margin: 4px 0;
+  font-size: 12px;
 `;
 
 const EditTaskDialog = () => {
@@ -279,8 +286,10 @@ const EditTaskDialog = () => {
   };
 
   const handleDelete = () => {
-    dispatch(deleteTask(task.id));
-    handleClose();
+    if (window.confirm("Are you sure? Deleting a task cannot be undone.")) {
+      dispatch(deleteTask(task.id));
+      handleClose();
+    }
   };
 
   const handleDescriptionClick = () => {
@@ -474,11 +483,21 @@ const EditTaskDialog = () => {
               font-size: 12px;
               font-weight: bold;
               color: ${TASK_G};
-              margin-bottom: 1rem;
+              margin-bottom: 2rem;
             `}
           >
             Delete task
           </Button>
+          <Text>
+            Updated {formatDistanceToNow(new Date(task.modified))} ago
+          </Text>
+          <Text
+            css={css`
+              margin-bottom: 1rem;
+            `}
+          >
+            Created {formatDistanceToNow(new Date(task.created))} ago
+          </Text>
         </Side>
       </Content>
     </Dialog>
